backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.3
python-dateutil==2.8.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
